Creating deltakit_core.api.paths
