+++ deltakit-explorer/tests/utils/test_config_file.py
@@ -4,13 +4,14 @@ from __future__ import annotations
 import os
 from pathlib import Path
 
+import deltakit_core.api.paths
 import pytest
 from deltakit_explorer._utils import _utils as utils
 
 
 def test_read_and_write_and_override_config_file(mocker):
     mocker.patch("deltakit_explorer._utils._utils.APP_NAME", "qec-testplorer")
-    config_file = utils.get_config_file_path()
+    config_file = deltakit_core.api.paths.get_config_file_path()
 
     # reset config file
     utils.override_persisted_variables({}, config_file)
@@ -40,7 +41,7 @@ def test_read_and_write_and_override_config_file(mocker):
 
 def test_variables_reading_to_environ(mocker):
     mocker.patch("deltakit_explorer._utils._utils.APP_NAME", "qec-testplorer")
-    config_file = utils.get_config_file_path()
+    config_file = deltakit_core.api.paths.get_config_file_path()
     utils.override_persisted_variables({"A": "B"}, config_file)
     utils.load_environment_variables_from_drive()
     assert os.environ.get("A") == "B"
@@ -60,7 +61,7 @@ def test_platform_specific_paths(mocker, platform, path):
         path = Path(os.getenv("XDG_CONFIG_HOME")) / "deltakit-explorer/.env"
     if platform == "win32" and os.getenv("APPDATA"):
         path = Path(os.getenv("APPDATA")) / "deltakit-explorer/.env"
-    assert utils.get_config_file_path() == path
+    assert deltakit_core.api.paths.get_config_file_path() == path
 
 
 @pytest.mark.parametrize(
@@ -70,31 +71,31 @@ def test_platform_specific_paths(mocker, platform, path):
 def test_platform_specific_paths_is_overridden(mocker, platform, envvar):
     os.environ[envvar] = str(Path.home() / "mock")
     mocker.patch("sys.platform", platform)
-    assert utils.get_config_file_path() == Path.home() / "mock/deltakit-explorer/.env"
+    assert deltakit_core.api.paths.get_config_file_path() == Path.home() / "mock/deltakit-explorer/.env"
 
 
 def test_error_creating_the_folder(mocker):
     mocker.patch("pathlib.Path.exists", return_value=False)
     mocker.patch("pathlib.Path.mkdir", side_effect=PermissionError(""))
-    assert utils.get_config_directory() == Path.cwd()
+    assert deltakit_core.api.paths.get_config_directory() == Path.cwd()
 
 
 def test_recreates_the_config_folder(mocker):
     mocker.patch("deltakit_explorer._utils._utils.APP_NAME", "qec-testplorer")
-    directory = utils.get_config_directory()
-    file = utils.get_config_file_path()
+    directory = deltakit_core.api.paths.get_config_directory()
+    file = deltakit_core.api.paths.get_config_file_path()
     if file.exists():
         Path.unlink(file)
     os.removedirs(directory)
     assert not directory.exists(), "Should have been deleted"
-    assert utils.get_config_directory().exists()
+    assert deltakit_core.api.paths.get_config_directory().exists()
     os.removedirs(directory)
 
 
 @pytest.mark.parametrize("content", ["faulty_abc", "faulty_abc=abc=abc"])
 def test_faulty_content(mocker, content):
     mocker.patch("deltakit_explorer._utils._utils.APP_NAME", "qec-testplorer")
-    file = utils.get_config_file_path()
+    file = deltakit_core.api.paths.get_config_file_path()
     with Path.open(file, "w", encoding="utf-8") as f:
         f.write(content)
     utils.load_environment_variables_from_drive()
